feat: move some constant lines to constant.tsx
